Finish moving relevant items from main to fishingView
