+++ src/main.rs
@@ -1,7 +1,7 @@
+extern crate rand;
 use bevy::window::EnabledButtons;
-use bevy::{prelude::*, window::PresentMode, sprite::{MaterialMesh2dBundle, Mesh2dHandle},};
+use bevy::{prelude::*, window::PresentMode};
 use rand::Rng;
-use std::time::Duration;
 
 mod physics;
 mod fish;
@@ -83,11 +83,6 @@ fn main() {
         .add_systems(Update, move_camera.after(move_player).run_if(in_state(FishingMode::Overworld)))
         .add_systems(Update, screen_edge_collision.after(move_player))
 
-            
-        .add_systems(Update, move_fish.run_if(in_state(FishingMode::Fishing)))
-        .add_systems(Update, fish_area_bobber.run_if(in_state(FishingMode::Fishing)))
-
-
         // Weather updates
         .add_systems(Update, update_weather)
         .add_systems(Update, update_weather_tint.after(update_weather))
@@ -105,8 +100,6 @@ fn main() {
         .run();
 }
 
-
-
 fn setup(
     mut commands: Commands,
     asset_server: Res<AssetServer>,
@@ -115,21 +108,13 @@ fn setup(
     mut materials: ResMut<Assets<ColorMaterial>>,
 ) {
     let mut rng = rand::thread_rng();
-
+    
     commands.spawn((
         Camera2dBundle::default(),
         Animation::new()
     ));
 
     commands.insert_resource(PlayerReturnPos {player_save_x: 0., player_save_y: 0.});
-
-    //commands.insert_resource(FishBoundsDir {change_x: Vec3::new(0.,0.,0.), change_y: Vec3::new(0.,0.,0.)});
-
-    commands.insert_resource(DirectionTimer{
-        // create the repeating timer
-        timer: Timer::new(Duration::from_secs(3), TimerMode::Repeating),
-    });
-    //let mut fish: HashMap<String, Species> = HashMap::new();
 
     
     //GRASS CODE V
@@ -236,7 +221,6 @@ fn setup(
     let player_layout_len = player_layout.textures.len();
     let player_layout_handle = texture_atlases.add(player_layout);
     let tree_sheet_handle: Handle<Image> = asset_server.load("tiles/tree.png"); 
-    let fish_bass_handle: Handle<Image> = asset_server.load("fish/bass.png");
 
     // MAP
     let map: Map = Map {
@@ -274,123 +258,6 @@ fn setup(
         },
         Animation::new()
     ));
-    commands.spawn((
-        SpriteBundle {
-            texture: fish_bass_handle.clone(),
-                sprite: Sprite {
-                custom_size: Some(Vec2::new(100.,100.)),
-                ..default()
-            },
-            transform: Transform {
-                translation: Vec3::new(FISHINGROOMX, FISHINGROOMY + 100., 0.),
-                ..default()
-            },
-            ..default()
-        },
-        BASS,
-        Fish {
-            id: 0,
-            is_caught: false,
-            is_alive: true,
-            touching_lure: false,
-            length: 8.0,
-            width: 5.0,
-            weight: 2.0,
-            age: 6.0,
-            hunger: 10.0
-        },
-        PhysicsObject {
-            mass: 2.0,
-            position: Vec3::new(FISHINGROOMX, FISHINGROOMY + 100., 0.),
-            rotation: Vec3::ZERO,
-            velocity: Vec3::ZERO,
-            forces: Forces::default()
-        }
-    ));
-    
-    //spawn example fish
-    //BEMMY
-    //BASS
-    let cool_fish_handle: Handle<Image> = asset_server.load("awesomeFishy.png");
-    commands.spawn((
-        SpriteBundle {
-            texture: cool_fish_handle.clone(),
-                sprite: Sprite {
-                custom_size: Some(Vec2::new(320.,180.)),
-                ..default()
-            },
-            transform: Transform {
-                translation: Vec3::new(FISHINGROOMX, FISHINGROOMY, 901.),
-                ..default()
-            },
-            ..default()
-        },
-        FishDetails {
-            name: "Bass1",
-            fish_id: 1,
-            length: rng.gen_range(4..8),
-            width: rng.gen_range(1..3),
-            weight: rng.gen_range(3..10),
-            time_of_day: (2,15),
-            weather: Weather::Sunny,
-            depth: (0,5),
-            //x, y, z
-            position: (8320, 3960),
-            change_x: Vec3::new(0.,0.,0.),
-            change_y: Vec3::new(0.,0.,0.),
-            //length, width, depth
-            bounds: (FISHINGROOMX as i32+100, FISHINGROOMY as i32 + 100),
-            hunger: 10.,
-            touching_lure: false,
-        },
-        InPond,
-        BASS,
-        Collision,
-    ));
-
-    //FISH BOX
-    commands.spawn((
-        SpriteBundle {
-            texture: cool_fish_handle.clone(),
-                sprite: Sprite {
-                custom_size: Some(Vec2::new(320.,180.)),
-                ..default()
-            },
-            transform: Transform {
-                translation: Vec3::new(FISHINGROOMX-40., FISHINGROOMY+40., 901.),
-                ..default()
-            },
-            ..default()
-        },
-        FishDetails {
-            name: "Cat1",
-            fish_id: 2,
-            length: rng.gen_range(5..12),
-            width: rng.gen_range(3..5),
-            weight: rng.gen_range(3..10),
-            time_of_day: (2,15),
-            weather: Weather::Rainy,
-            depth: (5,20),
-            //x, y, z
-            position: (8320, 3960),
-            change_x: Vec3::new(0.,0.,0.),
-            change_y: Vec3::new(0.,0.,0.),
-            //length, width, depth
-            bounds: (FISHINGROOMX as i32+100, FISHINGROOMY as i32 + 100),
-            hunger: 7.,
-            touching_lure: false,
-        },
-        InPond,
-        CATFISH,
-        Collision,
-    ));
-    /*
-    
-    
-        MAKE THE FISH STRUCT HERE :(
-    
-     */
-
     
     commands.spawn((
         SpriteBundle {
@@ -417,153 +284,9 @@ fn setup(
         GameDayTimer::new(10.),
     );
 
-
-    //let fishing_sheet_handle = asset_server.load("fishingView.png");
-
     //let grass_layout_len = grass_layout.textures.len();
-    
-    let fishing_sheet_handle: Handle<Image> = asset_server.load("fishingStuff/fishingView.png");
     //let tree_sheet_handle: Handle<Image> = asset_server.load("tiles/tree.png"); 
 
-    commands.spawn((
-        SpriteBundle {
-            texture: fishing_sheet_handle.clone(),
-                        sprite: Sprite {
-                        ..default()
-                    },
-            transform: Transform {
-                translation: Vec3::new(FISHINGROOMX, FISHINGROOMY, 900.),
-                ..default()
-            },
-            ..default()
-        },
-        
-    ));
-    
-    //powerbar view
-    let bar_sheet_handle = asset_server.load("fishingStuff/powerBar.png");
-    commands.spawn((
-        SpriteBundle {
-            texture: bar_sheet_handle.clone(),
-                        sprite: Sprite{
-                        ..default() 
-                        },
-            //where do I put it
-            transform: Transform {
-                translation: Vec3::new(FISHINGROOMX+575., FISHINGROOMY-308., 899.),
-                ..default()
-            },
-            ..default()
-        },
-        PowerBar
-    ));
-
-    let player_fishing_handle = asset_server.load("fishingStuff/backFishingSprite.png");
-    commands.spawn((
-        SpriteBundle {
-            texture: player_fishing_handle.clone(),
-                        sprite: Sprite{
-                        ..default() 
-                        },
-            //where do I put it
-            transform: Transform {
-                translation: Vec3::new(FISHINGROOMX-100., FISHINGROOMY-(WIN_H/2.)+50., 901.),
-                ..default()
-            },
-            ..default()
-        },
-    ));
-
-    let fishing_rod_handle = asset_server.load("fishingStuff/fishingRod.png");
-    commands.spawn((
-        SpriteBundle {
-            texture: fishing_rod_handle.clone(),
-                        sprite: Sprite{
-                        ..default() 
-                        },
-            //where do I put it
-            transform: Transform {
-                translation: Vec3::new(FISHINGROOMX-90., FISHINGROOMY-(WIN_H/2.)+100., 900.),
-                ..default()
-            },
-            ..default()
-        },
-        FishingRod {
-            length: 300.
-        }
-    ));
-
-    commands.spawn((
-        MaterialMesh2dBundle {
-            mesh: Mesh2dHandle(meshes.add(Rectangle::new(FishingLine::WIDTH, 0.0))),
-            material: materials.add(Color::hsl(100.,1., 1.)),
-            transform: Transform::from_xyz(FISHINGROOMX-90., FISHINGROOMY-(WIN_H/2.)+100.,   950.),
-            ..default()
-        },
-        FishingLine::default()
-    ));
-
-    let splashes_sheet_handle: Handle<Image> = asset_server.load("splashes/splashes.png");
-    let splash_layout = TextureAtlasLayout::from_grid(UVec2::new(100, 100), 3, 1, None, None);
-    let splash_layout_len = splash_layout.textures.len();
-    let splash_layout_handle = texture_atlases.add(splash_layout);
-    commands.spawn((
-        SpriteBundle {
-            texture: splashes_sheet_handle.clone(),
-            transform: Transform::from_xyz(FISHINGROOMX-90., FISHINGROOMY-(WIN_H/2.)+100.,   930.),
-            visibility: Visibility::Hidden,
-            ..default()
-        },
-        TextureAtlas {
-            layout: splash_layout_handle.clone(),
-            index: 0,
-        },
-        AnimationTimer::new(0.2), 
-        AnimationFrameCount(splash_layout_len), //number of different frames that we have
-        Splash::default(),
-        Animation::new()
-    ));
-
-    let waves_sheet_handle: Handle<Image> = asset_server.load("waves/waves.png");
-    let wave_layout = TextureAtlasLayout::from_grid(UVec2::new(100, 100), 4, 1, None, None);
-    let wave_layout_len = wave_layout.textures.len();
-    let wave_layout_handle = texture_atlases.add(wave_layout);
-    commands.spawn((
-        SpriteBundle {
-            texture: waves_sheet_handle.clone(),
-            transform: Transform::from_xyz(FISHINGROOMX-90., FISHINGROOMY-(WIN_H/2.)+100.,   930.),
-            visibility: Visibility::Hidden,
-            ..default()
-        },
-        TextureAtlas {
-            layout: wave_layout_handle.clone(),
-            index: 0,
-        },
-        //AnimationTimer::new(0.2), 
-        AnimationFrameCount(wave_layout_len), //number of different frames that we have
-        Wave,
-        //Animation::new()
-    ));
-
-    let bobber_handle = asset_server.load("fishingStuff/bobber.png");
-    commands.spawn((
-        SpriteBundle {
-            texture: bobber_handle.clone(),
-            sprite: Sprite {
-                custom_size: Some(Vec2::new(100.,100.)),
-                ..default()
-            },
-            transform: Transform {
-                translation: Vec3::new(FISHINGROOMX-90., FISHINGROOMY-(WIN_H/2.)+100.,   930.),
-                ..default()
-            },
-            visibility: Visibility::Hidden,
-            ..default()
-        },
-        Tile::BOBBER,
-        Collision,
-        Bobber::default(),
-    ));
     
     spawn_fishing_button(&mut commands, &asset_server);
     spawn_money_display(&mut commands, &asset_server);
